[REFACTORING]: Refactor scroll spy to separate file
